map앱 삭제후 posts와 결합. posts detail에 공구스팟,내위치 거리 추가
