+++ posts/models.py
@@ -30,3 +30,10 @@ class Spot(models.Model):
 
     def __str__(self):
         return self.spot
+
+class Search(models.Model):
+    address = models.CharField(max_length=200, null=True)
+    date = models.DateTimeField(auto_now_add=True)
+
+    def __str__(self):
+        return self.address
