+++ purchase_mate/settings.py
@@ -45,7 +45,6 @@ INSTALLED_APPS = [
     "allauth",
     "allauth.account",
     "allauth.socialaccount",
-    "map",
     "crispy_forms",
 ]
